requirements.txt
appdirs==1.4.4
asgiref==3.2.7
cfgv==3.1.0
distlib==0.3.1
Django==3.0.7
django-countries==6.1.2
djangorestframework==3.11.0
filelock==3.0.12
identify==1.4.20
importlib-metadata==1.7.0
importlib-resources==2.0.1
mysqlclient==1.4.6
nodeenv==1.4.0
pendulum==2.1.0
pre-commit==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2019.3
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
virtualenv==20.0.25
zipp==3.1.0
